feat: create MVP using serverless w/python
This is an MVP using serverless with the python-requirements and
step-function plugins. It can run a test plan such as the sample
one in the tests/fixtures.py file, and ensure it is cleaned up
properly.

It currently has several hacks in place, such as requiring the step
to provide cpu_units that max out the desired instance type so
that the scheduler ensures they are placed on unique nodes.

Closes #17